refactor: rename world/accessor -> world/rw and split methods to files
